+++ packages/plugin-couchdb/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.12](https://github.com/microfleet/core/compare/@microfleet/plugin-couchdb@2.2.11...@microfleet/plugin-couchdb@2.2.12) (2021-04-15)
+
+**Note:** Version bump only for package @microfleet/plugin-couchdb
+
+
+
+
+
 ## [2.2.11](https://github.com/microfleet/core/compare/@microfleet/plugin-couchdb@2.2.10...@microfleet/plugin-couchdb@2.2.11) (2021-02-19)
 
 **Note:** Version bump only for package @microfleet/plugin-couchdb